+++ packages/devtools_app_shared/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.1.1-wip
+* Update `package:dtd` to `^2.1.0`
+* Add `DTDManager.projectRoots` method.
+
 ## 0.1.0
 * Remove deprecated `background` and `onBackground` values for `lightColorScheme`
 and `darkColorScheme`.
@@ -6,6 +10,7 @@ and `darkColorScheme`.
 * Update readme to use `pub add` instead of explicit package version.
 * Update `package:dtd` to `^2.0.0`
 * Update `package:devtools_shared` to `^8.1.0`
+* Add `DTDManager.workspaceRoots` method.
 
 ## 0.0.10
 * Add `DTDManager` class and export from `service.dart`.